backend/requirements.txt
fastapi
uvicorn[standard]
openai==1.52.2
python-dotenv
httpx>=0.27.0